patch_tis bugfixes
If a mod introduces a new area which is a copy of one of the game's existing areas, and instead of copying the .tis to have the new name, it only edits it's .wed to point to the old .tis, the Widescreen mod fails to install. I've fixed it in the easiest way, which is by checking whether the .tis file exists and if it doesn't, read it's reference from the .wed. Redundant checks are there to prevent re-fixing the .tis file, and .tis files without .wed files are ignored to prevent enlargening water or lava overlays. -- Crevs Daak